Implement ML Kit OCR and refine data extraction architecture
- Updated the `Extractor` interface to return an `AppResult` and introduced `ExtractorError.MaskNotSetError` to handle missing configuration states.
- Implemented `MlkitOcrService` using Google ML Kit's Japanese text recognition, including logic to sort detected text blocks by vertical and horizontal coordinates.
- Modified the `OcrService` interface to accept `Bitmap` inputs and return a structured `OcrResult`.
- Integrated OCR and masking logic into `PayPayReceiptExtractor` to facilitate the extraction of expense data from receipt images.
- Refactored bitmap masking utility calls in `OCRViewModel` and `OCRMaskRatioAdjustView` to use `Double` precision for scaling ratios and updated their import paths.
- Added foundational parsing logic within the extraction flow to return `SentData` objects.

+++ app/src/main/java/gaku/original/myapplication/viewModel/ocr/OCRMaskRatioAdjustView.kt
@@ -10,9 +10,8 @@ import dagger.hilt.android.lifecycle.HiltViewModel
 import gaku.original.myapplication.data.Constants.Status.FuncStatus
 import gaku.original.myapplication.data.FuncResultWithData
 import gaku.original.myapplication.data.dataClass.SharedImageData
-import gaku.original.myapplication.data.repository.SharedPreferencesRepository
+import gaku.original.myapplication.data.extraction.extractor.maskBitmapTopLeftArea
 import gaku.original.myapplication.utility.loadBitmapFromUri
-import gaku.original.myapplication.utility.maskBitmapTopLeftArea
 import gaku.original.myapplication.viewModel.shared.SharedImageViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -71,8 +70,8 @@ class OCRMaskRatioAdjustViewModel @Inject constructor(
          */
         val maskedBitmap = maskBitmapTopLeftArea(
             source = bitmap,
-            widthPercent = _leftRatio.value,
-            heightPercent = _topRatio.value,
+            widthPercent = _leftRatio.value.toDouble(),
+            heightPercent = _topRatio.value.toDouble(),
         )
 
         return FuncResultWithData.Success(
